Add test for user show page
